Link app context with tenancy, remove app tenancy middleware
